+++ ecommerce-application/src/__tests__/RegistrationPage.test.tsx
@@ -1,0 +1,15 @@
+import { render } from '@testing-library/react';
+import { describe, it } from 'vitest';
+
+import { BrowserRouter } from 'react-router-dom';
+import { RegistrationPage } from '../components/Pages/RegistrationPage/RegistrationPage';
+
+describe('Renders RegistrationPage correctly', async () => {
+  it('Should render the RegistrationPage correctly', async () => {
+    render(
+      <BrowserRouter>
+        <RegistrationPage />
+      </BrowserRouter>,
+    );
+  });
+});
