+++ ecommerce-application/src/__tests__/Toast.test.tsx
@@ -1,0 +1,10 @@
+import { render } from '@testing-library/react';
+import { describe, it } from 'vitest';
+
+import { Toast } from '../components/UI/Toast/Toast';
+
+describe('Renders Toast correctly', async () => {
+  it('Should render the Toast correctly', async () => {
+    render(<Toast />);
+  });
+});
